Fix col names and prep data before db write

+++ sql/init_table_swell.postgres.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS swell (
     station INTEGER NOT NULL,
     date_utc DATE NOT NULL,
     time_utc TIMESTAMPZ NOT NULL,
-    hs_m REAL,
-    tp_s REAL,
-    dp_deg REAL,
-    ta_s REAL
+    hs_m DOUBLE PRECISION,
+    tp_s DOUBLE PRECISION,
+    dp_deg DOUBLE PRECISION,
+    ta_s DOUBLE PRECISION
 );
